+++ Entities/src/main/java/tw/waterball/judgegirl/entities/exam/ExamHasNotBeenStartedException.java
@@ -3,5 +3,5 @@ package tw.waterball.judgegirl.entities.exam;
 /**
  * @author - johnny850807@gmail.com (Waterball)
  */
-public class NoSubmissionQuotaException extends IllegalStateException {
+public class ExamHasNotBeenStartedException extends IllegalStateException {
 }
