+++ Exam-Service/src/main/java/tw/waterball/judgegirl/examservice/domain/repositories/ExamFilter.java
@@ -1,4 +1,4 @@
-package tw.waterball.judgegirl.examservice.repositories;
+package tw.waterball.judgegirl.examservice.domain.repositories;
 
 import lombok.Builder;
 import lombok.Getter;
